Cleared up the terminology around creating ports.
In Meep you can either open an existing port or create a new port for
other programs to open. The latter is called a virtual port in most
libraries, but in fact all ports are virtual. Even if there is a
physical device involved you will be communicating with a driver
through a virtual port.

The open vs. create port distinction will hopefully be a clearer mental model.

You can think of each port has having two ends. When you create an
output port it will appear in other programs as an input port. Thus,
messages you put into the input end will travel to the output end.

What's a little bit confusing is that messages go into the output end
and out of the input end. I will have to think about this a bit more.

I have probably introduced some bugs with this rewrite but I'll clean
that up later.